Restructure EMA-related namespaces for improved organization and update references accordingly.
