+++ app/Console/Commands/Ema/CreateEmaProducts.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace App\Console\Commands;
+namespace App\Console\Commands\Ema;
 
-use App\Jobs\DownloadEmaProducts;
+use App\Jobs\Ema\DownloadEmaProducts;
 use Illuminate\Console\Command;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Storage;
